+++ app/extend/helper.js
@@ -1,0 +1,15 @@
+'use strict';
+
+module.exports = {
+    parseInt(string) {
+        if (typeof string === 'number') {
+            return string;
+        }
+
+        if (!string) {
+            return string;
+        }
+
+        return parseInt(string, 10) || 0;
+    }
+};