+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## 0.0.3 (2018-08-31)
+## [0.0.4](http://icode.baidu.com/repos/baidu/hi-fe/node-demo/merge/v0.0.3...v0.0.4) (2018-09-05)
+
+
+### 新特性
+
+* add plugin egg-sequelize and mysql2 ([980f02a](http://icode.baidu.com/repos/baidu/hi-fe/node-demo/commits/980f02a)), 关闭卡片 [#17](http://newicafe.baidu.com/issue/node-best-practices-17/show) [#18](http://newicafe.baidu.com/issue/node-best-practices-18/show)
+* add validate plugin and add local & prod env ([ee19c8d](http://icode.baidu.com/repos/baidu/hi-fe/node-demo/commits/ee19c8d)), 关闭卡片 [#16](http://newicafe.baidu.com/issue/node-best-practices-16/show)
+
+
+## [0.0.3](http://icode.baidu.com/repos/baidu/hi-fe/node-demo/merge/1f646a8...v0.0.3) (2018-08-31)
 
 
 ### 新特性